requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
pdfplumber>=0.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
chromadb>=0.4.0
python-dotenv>=1.0.0
